requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
playwright==1.45.0
groq
pydantic>=2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
